If dotfiles called without arguments then update dotfiles locally

+++ .zshrc
@@ -68,5 +68,9 @@ $(prompt_char) "
 }
 
 function dotfiles {
-    ssh "$1" "if [ -d .git ]; then git pull; else git clone -n https://github.com/vslinko/dotfiles.git && mv dotfiles/.git . && rm -r dotfiles && git reset --hard; fi"
+    if [ $1 ]; then
+        ssh "$1" "if [ -d .git ]; then git pull; else git clone -n https://github.com/vslinko/dotfiles.git && mv dotfiles/.git . && rm -r dotfiles && git reset --hard; fi"
+    else
+        git pull && source .zshrc
+    fi
 }
